:recycle: create a namespace
and move initial set of data functions to new namespaced files.
