+++ bu-navigation.php
@@ -55,6 +55,8 @@ require_once BU_NAV_PLUGIN_DIR . '/includes/settings.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/library.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/class-tree-view.php';
 require_once BU_NAV_PLUGIN_DIR . '/includes/class-reorder.php';
+require_once BU_NAV_PLUGIN_DIR . '/includes/data-model.php';
+require_once BU_NAV_PLUGIN_DIR . '/includes/data-format.php';
 
 /**
  * Convenience class to wrap loading and init functions.
